requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via click
dnspython==2.7.0
    # via pymongo
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
narwhals==1.41.0
    # via altair
numpy==2.2.6
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   streamlit
pandas==2.3.0
    # via streamlit
pillow==11.2.1
    # via streamlit
protobuf==6.31.1
    # via streamlit
pyarrow==20.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pymongo==4.13.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   streamlit
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.45.1
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
typing-extensions==4.14.0
    # via
    #   altair
    #   referencing
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
watchdog==6.0.0
    # via streamlit
